Rename package configparser -> configfile & implement configfile.Find
