+++ configfile/parse_test.go
@@ -1,4 +1,4 @@
-package configparser_test
+package configfile_test
 
 import (
 	"net/url"
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/benchttp/runner/config"
-	"github.com/benchttp/runner/configparser"
+	"github.com/benchttp/runner/configfile"
 )
 
 const (
@@ -29,7 +29,7 @@ func TestParse(t *testing.T) {
 			expCfg := newExpConfig()
 			fname := path.Join(testdataConfigPath, "benchttp"+ext)
 
-			gotCfg, err := configparser.Parse(fname)
+			gotCfg, err := configfile.Parse(fname)
 			if err != nil {
 				// critical error, stop the test
 				t.Fatal(err)
